+++ src/main/java/com/exult/ProjectCisco/repository/FeatureXdeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FeatureXdeRepository extends JpaRepository<FeatureXde,Integer> {
+public interface FeatureXdeRepository extends JpaRepository<FeatureXde,Long> {
 
 }
