+++ src/main/java/com/exult/ProjectCisco/repository/MavenRepository.java
@@ -4,5 +4,5 @@ import com.exult.ProjectCisco.model.Maven;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-public interface MavenRepository extends JpaRepository<Maven,Integer> {
+public interface MavenRepository extends JpaRepository<Maven,Long> {
 }
